requirements.txt
numpy==2.2.4
python_igraph==0.11.8
